模块：自定义导航 修改：地图增加零点坐标轴显示，红色箭头所指为地图坐标X轴正向， 绿色箭头所指为Y轴正向。蓝色点为坐标零点。红色四边形箭头为机器人所在位置，其中，箭头方向为机器人的朝向。默认情况下，机器人开机的位置，即为地图零点位置。机器人朝向为X轴正向，为0度，角度逆时针增加。

+++ common/src/main/java/com/evertrend/tiger/common/bean/mapview/utils/GraphicalUtil.java
@@ -6,7 +6,7 @@ import android.graphics.Path;
 
 public class GraphicalUtil {
     public static void drawArrow(Canvas canvas, Paint paint, float fromX, float fromY, float toX, float toY,
-                          int heigth, int bottom) {
+                          float heigth, float bottom) {
         // heigth和bottom分别为三角形的高与底的一半,调节三角形大小
         canvas.drawLine(fromX, fromY, toX, toY, paint);
         float juli = (float) Math.sqrt((toX - fromX) * (toX - fromX)
